+++ models/dao/notification.go
@@ -4,7 +4,7 @@ import "time"
 
 type Notification struct {
 	NotifyID   int       `json:"notify_id" gorm:"column:notify_id;autoIncrement"`
-	UserID     int       `json:"user_id" gorm:"column:user_id;not null"`
+	UserID     int64     `json:"user_id" gorm:"column:user_id;not null"`
 	Type       string    `json:"type" gorm:"column:type;not null"`
 	SourceID   int       `json:"source_id" gorm:"column:source_id;not null"`
 	IsRead     bool      `json:"is_read" gorm:"column:is_read;default:false"`
